app-server/requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==4.2.16
django-cors-headers==4.7.0
django-filter==24.3
django-import-export==3.0.0
django-redis==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
numpy==1.23.5
openpyxl==3.0.10
packaging==24.1
pandas==1.5.3
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
redis==4.5.5
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2